# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Payment Reliability Lab — local demo API",
    "version": "1.0.0",
    "description": "Synthetic, loopback-only demo API. No production authentication is claimed. Amounts are exact integer minor units serialized as strings. Versioning: breaking changes get a new /v2 prefix. Clients must tolerate unknown response fields and unknown enum values (for example new transfer states or trace event types)."
  },
  "paths": {
    "/v1/transfers": {
      "post": {
        "summary": "Accept a transfer (idempotent)",
        "description": "202 means durably accepted locally, not paid externally. The same scoped key with the same asset, amount and destination replays the original body (header Idempotent-Replayed: true), also after restarts or settlement. `note` is ancillary: first write wins and it is excluded from the fingerprint. A request rejected before acceptance does not consume the key.",
        "parameters": [
          {
            "name": "X-Demo-Account",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Seeded demo account. A local identity stub, NOT authentication."
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[\\x21-\\x7e]{1,128}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTransfer"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Accepted (new or replayed)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Acceptance"
                }
              }
            }
          },
          "400": {
            "description": "INVALID_REQUEST",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UNKNOWN_DEMO_ACCOUNT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "IDEMPOTENCY_CONFLICT or INSUFFICIENT_FUNDS",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "PAYLOAD_TOO_LARGE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List this account's transfers",
        "parameters": [
          {
            "name": "X-Demo-Account",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Seeded demo account. A local identity stub, NOT authentication."
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Opaque cursor from a previous page."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 25
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or unknown demo account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Unknown or inaccessible transfer",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/transfers/{id}": {
      "get": {
        "summary": "Current local state, exact amount, provider reference and last provider observation with its age",
        "parameters": [
          {
            "name": "X-Demo-Account",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Seeded demo account. A local identity stub, NOT authentication."
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^tr_[0-9a-f]{20,32}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or unknown demo account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Unknown or inaccessible transfer",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/transfers/{id}/trace": {
      "get": {
        "summary": "Cursor-paginated append-only trace (stable IDs, per-transfer sequence, causation links)",
        "parameters": [
          {
            "name": "X-Demo-Account",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Seeded demo account. A local identity stub, NOT authentication."
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^tr_[0-9a-f]{20,32}$"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Opaque cursor from a previous page."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 25
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or unknown demo account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Unknown or inaccessible transfer",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/transfers/{id}/evidence": {
      "get": {
        "summary": "Linked journal batches, provider attempts, observations and webhook deliveries (bounded)",
        "parameters": [
          {
            "name": "X-Demo-Account",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Seeded demo account. A local identity stub, NOT authentication."
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^tr_[0-9a-f]{20,32}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or unknown demo account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Unknown or inaccessible transfer",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/transfers/{id}/invariants": {
      "get": {
        "summary": "Invariants I1-I8 as pass / fail / unknown with evidence IDs",
        "parameters": [
          {
            "name": "X-Demo-Account",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Seeded demo account. A local identity stub, NOT authentication."
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^tr_[0-9a-f]{20,32}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or unknown demo account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Unknown or inaccessible transfer",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/balances": {
      "get": {
        "summary": "Available, reserved and run clearing balances for the demo account",
        "parameters": [
          {
            "name": "X-Demo-Account",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Seeded demo account. A local identity stub, NOT authentication."
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or unknown demo account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Unknown or inaccessible transfer",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/exceptions": {
      "get": {
        "summary": "Unknown outcomes, conflicting observations and invariant failures (distinct kinds)",
        "parameters": [
          {
            "name": "X-Demo-Account",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Seeded demo account. A local identity stub, NOT authentication."
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Opaque cursor from a previous page."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 25
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["unknown_outcome", "conflicting_observation", "invariant_failure"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or unknown demo account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Unknown or inaccessible transfer",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/provider/webhooks": {
      "post": {
        "summary": "Signed provider webhook receiver",
        "description": "Header X-Provider-Signature: `t=<unix seconds>,v1=<hex HMAC-SHA256 of \"<t>.<raw body>\">`, 300 s tolerance. The event is durably stored before 202. An exact redelivery returns 202 with duplicate=true; a reused event ID with different content returns 409 and is kept as conflicting evidence.",
        "parameters": [
          {
            "name": "X-Provider-Signature",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebhookEnvelope"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Persisted (or verified duplicate)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "received": {
                      "type": "boolean"
                    },
                    "duplicate": {
                      "type": "boolean"
                    },
                    "deliveryId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "INVALID_REQUEST",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "INVALID_SIGNATURE or SIGNATURE_EXPIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "EVENT_ID_CONFLICT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "PAYLOAD_TOO_LARGE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "TEMPORARILY_UNAVAILABLE: not persisted, retry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreateTransfer": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "asset": {
            "type": "string",
            "const": "DEMO_USD"
          },
          "amountMinor": {
            "type": "string",
            "maxLength": 64
          },
          "destination": {
            "type": "string",
            "pattern": "^demo:[a-z0-9-]{1,48}$"
          },
          "note": {
            "type": "string",
            "maxLength": 512
          }
        },
        "required": ["asset", "amountMinor", "destination"],
        "additionalProperties": false
      },
...
      "WebhookEnvelope": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "eventId": {
            "type": "string",
            "pattern": "^evt_[0-9a-f]{20,32}$"
          },
          "type": {
            "type": "string",
            "const": "provider.transfer.updated"
          },
          "providerReference": {
            "type": "string",
            "pattern": "^pref_[0-9a-f]{20,32}$"
          },
          "status": {
            "type": "string",
            "minLength": 1,
            "maxLength": 32
          },
          "providerSequence": {
            "type": "integer",
            "minimum": 0,
            "maximum": 1000000
          },
          "occurredAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
          },
          "asset": {
            "type": "string",
            "minLength": 1,
            "maxLength": 16
          },
          "amountMinor": {
            "type": "string",
            "pattern": "^[1-9][0-9]{0,12}$"
          },
          "destination": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64
          },
          "finalNoEffect": {
            "type": "boolean"
          },
          "note": {
            "type": "string",
            "maxLength": 512
          }
        },
        "required": [
          "eventId",
          "type",
          "providerReference",
          "status",
          "providerSequence",
          "occurredAt",
          "asset",
          "amountMinor",
          "destination",
          "finalNoEffect"
        ],
        "additionalProperties": false
      },
      "Acceptance": {
        "type": "object",
        "required": ["transferId", "accepted", "statusUrl"],
        "properties": {
          "transferId": {
            "type": "string"
          },
          "accepted": {
            "const": true
          },
          "statusUrl": {
            "type": "string"
          }
        }
      },
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message", "requestId"],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "requestId": {
                "type": "string"
              }
            }
          }
        }
      }
    }
  }
}